Set up new repository and add README, license, etc files
This used to be a gh-pages branch of the main MacVim repository. We have
now moved to a separate repository
(https://github.com/macvim-dev/macvim-dev.github.io) which allows us to
serve the page from the root org level, and also allow us to nest the
docs project within it so that we can host/build the docs pages
independently and still have the URL work in a relative fashion to this
page (https://macvim-dev.github.io/docs and https://macvim.org/docs).

The old branch will have a new meta-tag redirect to point to this site
(GitHub Pages doesn't support generic server-side 3xx status code
redirect).