Add robots.txt and sitemap.txt
The robots.txt file points to both a top-level sitemap and one generated
by the docs repo which links to all the generated docs files. The
top-level sitemap.txt only has a single root page since that's all we
have. Don't use XML sitemaps for now as there really isn't a need for
it.