Add "What's New" page and format release notes better
Add a new "What's New" page that will display the release notes in HTML
hosted from MacVim website instead of needing to go to GitHub releases
page. This will be used within MacVim to show a built-in "What's New"
page whenever the user has updated to a new version. It's a static site
that has the 4 latest releases (via Jekyll templating) with only the
latest release visible. MacVim will use URL parameters to specify which
version it wants to see (e.g. if a user upgraded from 176 to 178, the
page will show 177 and 178's release notes) and JavaScript on the page
will hide/show the relevant HTML elements.

Also, refactor release notes to be included and version controlled
instead of just being a part of the appcast XML. As part of this, commit
some recent historical release notes to populate the list and make them
easier to access.

Also refactor out the CSS styles for them and make the header looks
better with a line separator like GitHub Release's styling.