Add table of contents to What's New page
Include the relevant styles and a show/hide toggle button on the top
right (CSS only). Only show the TOC when the window is wide enough. If
there are multiple release notes shown, will automatically group them
together into a single TOC in script.

Also add a new "all" parameter that could be passed to the URL parameter
which is easier than having to type `?from=1&to=9999`.